test(delegate_to): add specs for ApplyStubToTrackDelegations and move to matcher namespace
