+++ lib/convenient_service/rspec/matchers/custom/delegate_to/entities.rb
@@ -1,7 +1,3 @@
 # frozen_string_literal: true
 
-require_relative "entities/chainings"
-require_relative "entities/chainings_collection"
-require_relative "entities/delegation"
 require_relative "entities/matcher"
-require_relative "entities/presenter"
