--- lib/convenient_service/rspec/matchers/custom/delegate_to/entities/chainings/with_calling_original.rb
@@ -1,34 +1,0 @@
-# frozen_string_literal: true
-
-module ConvenientService
-  module RSpec
-    module Matchers
-      module Custom
-        class DelegateTo
-          module Entities
-            module Chainings
-              class WithCallingOriginal < Chainings::Base
-                ##
-                # @param block_expectation_value [Object]
-                # @return [Boolean]
-                #
-                def matches?(block_expectation_value)
-                  super
-
-                  true
-                end
-
-                ##
-                # @return [Boolean]
-                #
-                def should_call_original?
-                  true
-                end
-              end
-            end
-          end
-        end
-      end
-    end
-  end
-end
