+++ lib/convenient_service/rspec/matchers/custom/delegate_to/entities/matcher/commands.rb
