+++ lib/convenient_service/rspec/matchers/custom/delegate_to/entities/matcher/entities.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require_relative "entities/chainings"
+require_relative "entities/chainings_collection"
+require_relative "entities/delegation"
+require_relative "entities/presenter"
