+++ lib/convenient_service/rspec/matchers/custom/delegate_to/entities/matcher/entities/chainings.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require_relative "chainings/base"
+require_relative "chainings/with_any_arguments"
+require_relative "chainings/with_concrete_arguments"
+require_relative "chainings/without_arguments"
+require_relative "chainings/return_its_value"
+require_relative "chainings/with_calling_original"
+require_relative "chainings/without_calling_original"
