+++ lib/convenient_service/rspec/matchers/custom/delegate_to/entities/matcher/entities/chainings/with_any_arguments.rb
@@ -1,0 +1,52 @@
+# frozen_string_literal: true
+
+module ConvenientService
+  module RSpec
+    module Matchers
+      module Custom
+        class DelegateTo
+          module Entities
+            class Matcher
+              module Entities
+                module Chainings
+                  class WithAnyArguments < Chainings::Base
+                    ##
+                    # @return [void]
+                    #
+                    def apply_stubs!
+                      Commands::ApplyStubToTrackDelegations.call(matcher: matcher)
+                    end
+
+                    ##
+                    # @param block_expectation_value [Object]
+                    # @return [Boolean]
+                    #
+                    def matches?(block_expectation_value)
+                      super
+
+                      matcher.delegations.any?
+                    end
+
+                    ##
+                    # @return [String]
+                    #
+                    def failure_message
+                      "expected `#{matcher.printable_block_expectation}` to delegate to `#{matcher.printable_method}` with any arguments at least once (no arguments is also valid), but it didn't."
+                    end
+
+                    ##
+                    # @return [String]
+                    #
+                    def failure_message_when_negated
+                      "expected `#{matcher.printable_block_expectation}` NOT to delegate to `#{matcher.printable_method}` with any arguments at least once (no arguments is also valid), but it did."
+                    end
+                  end
+                end
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+end
